% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ggPm.At.longroot.longitudinal}
\alias{ggPm.At.longroot.longitudinal}
\title{ggPm.At.longroot.longitudinal}
\usage{
data(ggPm.At.longroot.longitudinal)
}
\description{
Longitudinal section of Arabidopsis thaliana root
}
