% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ggPm.At.root.crosssection.d}
\alias{ggPm.At.root.crosssection.d}
\title{ggPm.At.root.crosssection.d}
\usage{
data(ggPm.At.root.crosssection.d)
}
\description{
Cross section of Arabidopsis thaliana root in the differentiation zone
}
