% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ggPm.At.root.crosssection.e1}
\alias{ggPm.At.root.crosssection.e1}
\title{ggPm.At.root.crosssection.e1}
\usage{
data(ggPm.At.root.crosssection.e1)
}
\description{
Cross section of Arabidopsis thaliana root in the early elongation zone
}
