% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ggPm.At.root.crosssection.e2}
\alias{ggPm.At.root.crosssection.e2}
\title{ggPm.At.root.crosssection.e2}
\usage{
data(ggPm.At.root.crosssection.e2)
}
\description{
Cross section of Arabidopsis thaliana root in the late elongation zone
}
