% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ggPm.At.root.crosssection.m1}
\alias{ggPm.At.root.crosssection.m1}
\title{ggPm.At.root.crosssection.m1}
\usage{
data(ggPm.At.root.crosssection.m1)
}
\description{
Cross section of Arabidopsis thaliana root in the early meristem zone
}
