% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ggPm.At.root.crosssection.t}
\alias{ggPm.At.root.crosssection.t}
\title{ggPm.At.root.crosssection.t}
\usage{
data(ggPm.At.root.crosssection.t)
}
\description{
Cross section of Arabidopsis thaliana root in the transition zone
}
