% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_function.R
\name{generate_common_palette}
\alias{generate_common_palette}
\title{Generate a Common Palette}
\usage{
generate_common_palette(Group_name, ..., color_palette = scales::hue_pal())
}
\arguments{
\item{Group_name}{Character. Name of the group column.}

\item{...}{Data frames to be included in the palette generation.}

\item{color_palette}{Function. A color palette function (default is \code{scales::hue_pal}).}
}
\value{
A named vector of colors.
}
\description{
Creates a consistent color palette for a given grouping variable across datasets.
}
