% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggRootCellAtlas_annotation.R
\name{ggRootCellAtlas_annotation}
\alias{ggRootCellAtlas_annotation}
\title{Generate Root Cell Atlas Annotation Visualizations}
\usage{
ggRootCellAtlas_annotation(Group_name)
}
\arguments{
\item{Group_name}{Character. Name of the column in the datasets that represents
the group annotations to be visualized.}
}
\value{
A ggplot2 object representing the combined annotated plots, with each plot
corresponding to a specific Arabidopsis thaliana root section and
a shared legend for group annotations.
}
\description{
This function creates annotated visualizations for various root sections.
It uses a shared color palette to represent group annotations and combines multiple
plots into a single composite visualization.
}
\section{Example usage}{
ggRootCellAtlas_annotation("Zones")
ggRootCellAtlas_annotation("Sections")
ggRootCellAtlas_annotation("SubCellTypes")
ggRootCellAtlas_annotation("CellTypes")
ggRootCellAtlas_annotation("TissueTypes")
ggRootCellAtlas_annotation("TissueSubTypes")
ggRootCellAtlas_annotation("Atlas")
ggRootCellAtlas_annotation("Atlas_reduce")
}

