% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggRootCellAtlas_expression.R
\name{ggRootCellAtlas_expression}
\alias{ggRootCellAtlas_expression}
\title{Generate Root Cell Atlas Gene Expression Heatmaps}
\usage{
ggRootCellAtlas_expression(
  avg_exp,
  Gene,
  c1 = NA,
  c2 = NA,
  Annotation = "Atlas"
)
}
\arguments{
\item{avg_exp}{A data frame containing average gene expression data. Columns should represent
different atlas regions, and rows should represent genes.}

\item{Gene}{A string specifying the gene for which expression or other feature data should be visualized.}

\item{c1}{Numeric value specifying the minimum value for the color gradient. Defaults to 0.}

\item{c2}{Numeric value specifying the maximum value for the color gradient. Defaults to the
maximum gene expression across all datasets.}

\item{Annotation}{Character. Name of the column to annotate that corresponds to Idents of Seurat object.
Defaults to the "Atlas" Idents}
}
\value{
A \code{ggplot2} object representing the combined heatmaps, with each heatmap corresponding
to a specific atlas dataset and a shared legend indicating the gene expression levels.
}
\description{
This function creates a series of heatmaps representing the expression of a specified gene
or oother cell fiature across various root sections. It combines the heatmaps into a single composite
visualization using a shared color gradient.
}
\section{Example usage}{
Seurat.object<-SetIdent(Seurat.object,value = "AnnotationName")
avg_exp <- AverageExpression(Seurat.object,assays = 'RNA',slot = 'data')
ggRootCellAtlas_expression(avg_exp, "GeneID", 1, 10, "AnnotationName")

ggRootCellAtlas_expression(avg_exp, Gene = "GeneID")
}

